+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Change Log
 
+## [v1.0.3](https://github.com/bvanhoekelen/performance/tree/v1.0.3) - 2017-04-16
+### Added
+- Config system
+- Config item: CONSOLE_LIVE
+
+### Changes
+- Text by 'max execution time' by 0 to unlimited
+
+### Remove
+- Performance ascii art. Pull request form nunomaduro.
+
 ## [v1.0.2](https://github.com/bvanhoekelen/performance/tree/v1.0.2) - 2017-03-23
 ### Fixed
 - Format to seconds